% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JapaneseVowels.R
\docType{data}
\name{JapaneseVowels}
\alias{JapaneseVowels}
\title{JapaneseVowels}
\format{
A \code{list} with two elements, which are:
\describe{
\item{\code{data}}{A list with 640 MTS.}
\item{\code{classes}}{A numeric vector indicating the corresponding classes associated with the elements in \code{data}.}
}
}
\usage{
data(JapaneseVowels)
}
\description{
Multivariate time series (MTS) indicating voice recordings of nine Japanese male speakers saying the vowels 'a' and 'e'.
}
\details{
Each element in \code{data} is a matrix formed by 29 rows (time points) indicating time recordings and 12 columns
(variables) indicating modified raw recordings. The first 270 elements correspond to the training set, whereas the last 370 elements
correspond to the test set. The numeric vector \code{classes} is formed by integers from 1 to 9, indicating that there are 9
different classes in the database. Each class is associated with a different speaker.
For more information, see \insertCite{bagnall2018uea;textual}{mlmts}.
Run "install.packages("ueadata1", repos="https://anloor7.github.io/drat")"
to access this dataset and use the syntax "ueadata1::JapaneseVowels".
}
\references{
{

  \insertRef{bagnall2018uea}{mlmts}

  \insertRef{ruiz2021great}{mlmts}

  \insertRef{bagnallweb}{mlmts}

}
}
\keyword{datasets}
