% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-mlmts.R
\docType{package}
\name{mlmts}
\alias{mlmts}
\title{mlmts: Machine Learning Algorithms for Multivariate Time Series.}
\description{
mlmts provides an implementation of several machine learning algorithms for
multivariate time series. The package includes functions allowing the
execution of clustering, classification or outlier detection methods,
among others. It also incorporates a collection of multivariate time
series datasets which can be used to analyse the performance of new
proposed algorithms. Practitioners from a broad variety of fields could
benefit from the general framework provided by mlmts.
}
