% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mts_plot.R
\encoding{UTF-8}
\name{mts_plot}
\alias{mts_plot}
\title{Constructs a plot of a MTS}
\usage{
mts_plot(series, title = "")
}
\arguments{
\item{series}{A MTS (numerical matrix).}

\item{title}{Title for the plot (string). Default corresponds to no title.}
}
\value{
The corresponding plot.
}
\description{
\code{mts_plot} constructs a plot of a MTS. Each univariate series comprising
the MTS object is displayed in a different colour.
}
\details{
Given a MTS, the function constructs the corresponding plot, in which a
different colour is used for each univariate series comprising the
MTS object. Therefore, the MTS is represented as a collection of univariate
series in a single graph.
}
\examples{
mts_plot(BasicMotions$data[[1]]) # Represents the first MTS in dataset
# BasicMotions
}
\author{
Ángel López-Oriona, José A. Vilar
}
