% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEMS_SF_2.R
\docType{data}
\name{PEMS_SF_2}
\alias{PEMS_SF_2}
\title{PEMS_SF_2}
\format{
A \code{list} with two elements, which are:
\describe{
\item{\code{data}}{A list with 220 MTS.}
\item{\code{classes}}{A numeric vector indicating the corresponding classes associated with the elements in \code{data}.}
}
}
\usage{
data(PEMS_SF_2)
}
\description{
Multivariate time series (MTS) indicating occupancy rate of different car lanes.
}
\details{
Each element in \code{data} is a matrix formed by 144 rows (time points) indicating minutes and 3 columns (variables) indicating sensors.
The last 220 elements
of the whole dataset are stored here. The last 173 elements of this dataset pertain to the test set. The numeric vector \code{classes} is formed
by integers from 1 to 7, indicating that there are 7 different classes in the database. Each class is associated with a different
day of the week. For more information, see \insertCite{bagnall2018uea;textual}{mlmts}.
Run "install.packages("ueadata8", repos="https://anloor7.github.io/drat")"
to access this dataset and use the syntax "ueadata8::PEMS_SF_2".
}
\references{
{

  \insertRef{bagnall2018uea}{mlmts}

  \insertRef{ruiz2021great}{mlmts}

  \insertRef{bagnallweb}{mlmts}

}
}
\keyword{datasets}
