% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelfRegulationSCP1.R
\docType{data}
\name{SelfRegulationSCP1}
\alias{SelfRegulationSCP1}
\title{SelfRegulationSCP1}
\format{
A \code{list} with two elements, which are:
\describe{
\item{\code{data}}{A list with 561 MTS.}
\item{\code{classes}}{A numeric vector indicating the corresponding classes associated with the elements in \code{data}.}
}
}
\usage{
data(SelfRegulationSCP1)
}
\description{
Multivariate time series (MTS) taken from a healthy subject asked to move a cursor up and down on a computer screen while
his cortical potentials were taken.
}
\details{
Each element in \code{data} is a matrix formed by 896 rows (time points) indicating time recordings over an interval of 3.5 seconds
and 6 columns (variables) indicating EEG channel. The first 268 elements correspond to the training set, whereas the last 293 elements
correspond to the test set. The numeric vector \code{classes} is formed by integers from 1 to 2, indicating that there are 2
different classes in the database. Each class is associated with the label 'negativity' (downward movement of the cursor) or 'positivity'
(upward movement of the cursor). For more information, see \insertCite{bagnall2018uea;textual}{mlmts}.
Run "install.packages("ueadata2", repos="https://anloor7.github.io/drat")"
to access this dataset and use the syntax "ueadata2::SelfRegulationSCP1".
}
\references{
{

  \insertRef{bagnall2018uea}{mlmts}

  \insertRef{ruiz2021great}{mlmts}

  \insertRef{bagnallweb}{mlmts}

}
}
\keyword{datasets}
