% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandWalkJump.R
\docType{data}
\name{StandWalkJump}
\alias{StandWalkJump}
\title{StandWalkJump}
\format{
A \code{list} with two elements, which are:
\describe{
\item{\code{data}}{A list with 27 MTS.}
\item{\code{classes}}{A numeric vector indicating the corresponding classes associated with the elements in \code{data}.}
}
}
\usage{
data(StandWalkJump)
}
\description{
Multivariate time series (MTS) involving short duration ECG signals recorded from a healthy 25-year-old male performing different
physical activities
}
\details{
Each element in \code{data} is a matrix formed by 2500 rows (time points) indicating readings in a spectrogram and 4 columns
(variables) indicating frequency band from the spectrogram. The first 12 elements correspond to the training set, whereas the last 15 elements
correspond to the test set. The numeric vector \code{classes} is formed by integers from 1 to 3, indicating that there are 3
different classes in the database. Each class is associated with the label 'standing', 'walking' or 'jumping'.
For more information, see \insertCite{bagnall2018uea;textual}{mlmts}.
}
\references{
{

  \insertRef{bagnall2018uea}{mlmts}

  \insertRef{ruiz2021great}{mlmts}

  \insertRef{bagnallweb}{mlmts}

}
}
\keyword{datasets}
